requirements.txt
psycopg2-binary>=2.9.7
django>=2.2.5
django-cors-headers>=3.10.1
django-filter>=21.1
djangorestframework>=3.12
markdown>=3.3.7
sqlparse>=0.4.4
numpy>=1.19.5
openpyxl>=3.1.2
websocket-client>=1.3.1
Jinja2>=3.0.3
itsdangerous>=2.0.1
pandas>=1.1.5
py2neo>=2021.2.3
uwsgi>=2.0.21
websockets>=9.1
